# Conflicts:
#	package.json
#	src/text-buffer.coffee

+++ spec/text-buffer-spec.coffee
@@ -11,6 +11,10 @@ SampleText = fs.readFileSync(join(__dirname, 'fixtures', 'sample.js'), 'utf8')
 describe "TextBuffer", ->
   buffer = null
 
+  beforeEach ->
+    # When running specs in Atom, setTimeout is spied on by default.
+    jasmine.useRealClock?()
+
   afterEach ->
     buffer = null
 
@@ -45,6 +49,12 @@ describe "TextBuffer", ->
       expect(buffer.lineEndingForRow(0)).toBe '\n'
       expect(buffer.lineForRow(1)).toBe ''
       expect(buffer.lineEndingForRow(1)).toBe ''
+
+    it "automatically assigns a unique identifier to new buffers", ->
+      bufferIds = [0..16].map(-> new TextBuffer().getId())
+      uniqueBufferIds = new Set(bufferIds)
+
+      expect(uniqueBufferIds.size).toBe(bufferIds.length)
 
     describe "when a file path is given", ->
       [filePath] = []
@@ -604,7 +614,8 @@ describe "TextBuffer", ->
           buffer.append("three\n")
           buffer.append("four")
 
-        buffer.markRange([[0, 1], [2, 3]], a: 'b', maintainHistory: true)
+        historyLayer = buffer.addMarkerLayer(maintainHistory: true)
+        historyLayer.markRange([[0, 1], [2, 3]], a: 'b')
         result = buffer.groupChangesSinceCheckpoint(checkpoint)
 
         expect(result).toBe true
@@ -626,7 +637,7 @@ describe "TextBuffer", ->
           four
         """
 
-        [marker] = buffer.getMarkers()
+        [marker] = historyLayer.getMarkers()
         expect(marker.getRange()).toEqual [[0, 1], [2, 3]]
         expect(marker.getProperties()).toEqual {a: 'b'}
 
@@ -807,26 +818,27 @@ describe "TextBuffer", ->
       expect(buffer.positionForCharacterIndex(20)).toEqual [3, 5]
 
   describe "serialization", ->
-    expectSameMarkers = (buffer1, buffer2) ->
-      markers1 = buffer1.getMarkers().sort (a, b) -> a.compare(b)
-      markers2 = buffer2.getMarkers().sort (a, b) -> a.compare(b)
+    expectSameMarkers = (left, right) ->
+      markers1 = left.getMarkers().sort (a, b) -> a.compare(b)
+      markers2 = right.getMarkers().sort (a, b) -> a.compare(b)
       expect(markers1.length).toBe markers2.length
       for marker1, i in markers1
         expect(marker1).toEqual(markers2[i])
       return
 
-    it "can serialize / deserialize the buffer along with its history and markers", ->
+    it "can serialize / deserialize the buffer along with its history, marker layers, and markers", ->
       bufferA = new TextBuffer(text: "hello\nworld\r\nhow are you doing?")
       bufferA.createCheckpoint()
       bufferA.setTextInRange([[0, 5], [0, 5]], " there")
       bufferA.transact -> bufferA.setTextInRange([[1, 0], [1, 5]], "friend")
-      bufferA.markRange([[0, 6], [0, 8]], reversed: true, foo: 1)
+      layerA = bufferA.addMarkerLayer(maintainHistory: true)
+      layerA.markRange([[0, 6], [0, 8]], reversed: true, foo: 1)
       marker2A = bufferA.markPosition([2, 2], bar: 2)
       bufferA.transact ->
         bufferA.setTextInRange([[1, 0], [1, 0]], "good ")
         bufferA.append("?")
         marker2A.setProperties(bar: 3, baz: 4)
-      bufferA.markRange([[0, 4], [0, 5]], invalidate: 'inside')
+      layerA.markRange([[0, 4], [0, 5]], invalidate: 'inside')
       bufferA.setTextInRange([[0, 5], [0, 5]], "oo")
       bufferA.undo()
 
@@ -862,13 +874,33 @@ describe "TextBuffer", ->
       expectSameMarkers(bufferB, bufferA)
 
       # Accounts for deserialized markers when selecting the next marker's id
-      marker3A = bufferA.markRange([[0, 1], [2, 3]])
-      marker3B = bufferB.markRange([[0, 1], [2, 3]])
+      marker3A = layerA.markRange([[0, 1], [2, 3]])
+      marker3B = bufferB.getMarkerLayer(layerA.id).markRange([[0, 1], [2, 3]])
       expect(marker3B.id).toBe marker3A.id
 
       # Doesn't try to reload the buffer since it has no file.
       waits(50)
       runs -> expect(bufferB.getText()).toBe "hello\nworld\r\nhow are you doing?"
+
+    it "serializes / deserializes the buffer's custom marker layers", ->
+      bufferA = new TextBuffer("abcdefghijklmnopqrstuvwxyz")
+
+      layer1A = bufferA.addMarkerLayer()
+      layer2A = bufferA.addMarkerLayer(maintainHistory: true)
+
+      layer1A.markRange([[0, 1], [0, 2]])
+      layer1A.markRange([[0, 3], [0, 4]])
+
+      layer2A.markRange([[0, 5], [0, 6]])
+      layer2A.markRange([[0, 7], [0, 8]])
+
+      bufferB = TextBuffer.deserialize(JSON.parse(JSON.stringify(bufferA.serialize())))
+      layer1B = bufferB.getMarkerLayer(layer1A.id)
+      layer2B = bufferB.getMarkerLayer(layer2A.id)
+      expect(layer2B.maintainHistory).toBe true
+
+      expectSameMarkers(layer1A, layer1B)
+      expectSameMarkers(layer2A, layer2B)
 
     it "doesn't serialize markers with the 'persistent' option set to false", ->
       bufferA = new TextBuffer(text: "hello\nworld\r\nhow are you doing?")
@@ -878,6 +910,12 @@ describe "TextBuffer", ->
       bufferB = TextBuffer.deserialize(bufferA.serialize())
       expect(bufferB.getMarker(marker1A.id)).toBeUndefined()
       expect(bufferB.getMarker(marker2A.id)).toBeDefined()
+
+    it "serializes / deserializes the buffer's unique identifier", ->
+      bufferA = new TextBuffer()
+      bufferB = TextBuffer.deserialize(JSON.parse(JSON.stringify(bufferA.serialize())))
+
+      expect(bufferB.getId()).toEqual(bufferA.getId())
 
     describe "when the buffer has a path", ->
       [filePath, buffer2] = []
@@ -900,17 +938,26 @@ describe "TextBuffer", ->
           expect(buffer.isModified()).toBeFalsy()
           buffer2 = buffer.testSerialization()
 
+          buffer2ModifiedEvents = []
+          buffer2.onDidChangeModified (value) -> buffer2ModifiedEvents.push(value)
+
           waitsFor ->
             buffer2.loaded
 
           runs ->
             expect(buffer2.isModified()).toBeFalsy()
-            expect(buffer2.getPath()).toBe(buffer.getPath())
-            expect(buffer2.getText()).toBe(buffer.getText())
+            expect(buffer2ModifiedEvents).toEqual [false]
+            expect(buffer2.getPath()).toBe(filePath)
+            expect(buffer2.getText()).toBe('words!')
 
             buffer.undo()
             buffer2.undo()
-            expect(buffer2.getText()).toBe(buffer.getText())
+
+          waits(buffer.stoppedChangingDelay)
+
+          runs ->
+            expect(buffer2.getText()).toBe('words')
+            expect(buffer2ModifiedEvents).toEqual [false, true]
 
       describe "when the serialized buffer had unsaved changes", ->
         describe "when the disk contents were changed since serialization", ->
@@ -1013,6 +1060,9 @@ describe "TextBuffer", ->
       expect(eventHandler).toHaveBeenCalledWith(newPath)
 
     it "notifies observers when the buffer's file is moved", ->
+      # FIXME: This doesn't pass on Linux
+      return if process.platform in ['linux', 'win32']
+
       fs.removeSync(newPath)
       fs.moveSync(filePath, newPath)
 
@@ -1165,6 +1215,9 @@ describe "TextBuffer", ->
           deleteHandler.callCount > 0
 
       it "retains its path and reports the buffer as not modified", ->
+        # FIXME: This doesn't pass on Linux
+        return if process.platform is 'linux'
+
         expect(bufferToDelete.getPath()).toBe filePath
         expect(bufferToDelete.isModified()).toBeFalsy()
 
@@ -1330,26 +1383,19 @@ describe "TextBuffer", ->
 
     it "returns false for an empty buffer with no path", ->
       buffer.destroy()
-      buffer = new TextBuffer({load: true})
-
-      waitsFor ->
-        buffer.loaded
-
-      runs ->
-        expect(buffer.isModified()).toBeFalsy()
+      buffer = new TextBuffer()
+      expect(buffer.isModified()).toBeFalsy()
+      buffer.append('hello')
+      expect(buffer.isModified()).toBeTruthy()
 
     it "returns true for a non-empty buffer with no path", ->
       buffer.destroy()
-      buffer = new TextBuffer({load: true})
-
-      waitsFor ->
-        buffer.loaded
-
-      runs ->
-        buffer.setText('a')
-        expect(buffer.isModified()).toBeTruthy()
-        buffer.setText('\n')
-        expect(buffer.isModified()).toBeTruthy()
+      buffer = new TextBuffer({text: 'something'})
+      expect(buffer.isModified()).toBeTruthy()
+      buffer.append('a')
+      expect(buffer.isModified()).toBeTruthy()
+      buffer.setText('')
+      expect(buffer.isModified()).toBeFalsy()
 
     it "returns false until the buffer is fully loaded", ->
       buffer.destroy()
@@ -1716,7 +1762,7 @@ describe "TextBuffer", ->
           saveBuffer.setText "hi"
           expect(-> saveBuffer.save()).toThrow()
 
-  describe "reload()", ->
+  describe "::reload()", ->
     it "reloads current text from disk and clears any conflicts", ->
       filePath = require.resolve('./fixtures/sample.js')
       fileContents = fs.readFileSync(filePath, 'utf8')
@@ -2168,7 +2214,7 @@ describe "TextBuffer", ->
 
     describe "when called with a random range", ->
       it "returns the same results as ::scanInRange, but in the opposite order", ->
-        for i in [1...10]
+        for i in [1...50]
           seed = Date.now()
           random = new Random(seed)
 
@@ -2186,12 +2232,31 @@ describe "TextBuffer", ->
             /.{5}/g
           ][random(4)]
 
-          forwardRanges = []
-          backwardRanges = []
-          buffer.scanInRange regex, range, ({range, matchText}) -> forwardRanges.push({range, matchText})
-          buffer.backwardsScanInRange regex, range, ({range, matchText}) -> backwardRanges.unshift({range, matchText})
+          if random(2) > 0
+            forwardRanges = []
+            backwardRanges = []
+            forwardMatches = []
+            backwardMatches = []
 
-          expect(backwardRanges).toEqual(forwardRanges, "Seed: #{seed}")
+            buffer.scanInRange regex, range, ({range, matchText}) ->
+              forwardMatches.push(matchText)
+              forwardRanges.push(range)
+
+            buffer.backwardsScanInRange regex, range, ({range, matchText}) ->
+              backwardMatches.unshift(matchText)
+              backwardRanges.unshift(range)
+
+            expect(backwardRanges).toEqual(forwardRanges, "Seed: #{seed}")
+            expect(backwardMatches).toEqual(forwardMatches, "Seed: #{seed}")
+          else
+            referenceBuffer = new TextBuffer(text: buffer.getText())
+            referenceBuffer.scanInRange regex, range, ({matchText, replace}) ->
+              replace(matchText + '.')
+
+            buffer.backwardsScanInRange regex, range, ({matchText, replace}) ->
+              replace(matchText + '.')
+
+            expect(buffer.getText()).toBe(referenceBuffer.getText(), "Seed: #{seed}")
 
   describe "::characterIndexForPosition(position)", ->
     beforeEach ->
